+++ bluepysnap/bbp.py
@@ -19,6 +19,9 @@
 
 from bluepysnap.sonata_constants import DYNAMICS_PREFIX, Edge, Node
 
+NODE_TYPES = {'biophysical', 'virtual', 'astrocyte', 'single_compartment', 'point_neuron'}
+EDGE_TYPES = {'chemical', 'electrical', 'synapse_astrocyte', 'endfoot'}
+
 
 class Cell(Node):
     """Cell property names."""