+++ CHANGELOG.rst
@@ -1,6 +1,20 @@
 Changelog
 =========
 
+Version v0.13.0
+---------------
+
+New Features
+~~~~~~~~~~~~~
+- Node/edge populations are now supported in config
+- Population type available in NodePopulation/EdgePopulation
+- Population config (if given) overwrites the "components" config for that population
+- Alternate morphology directories (.h5, .asc) are now supported
+
+Improvements
+~~~~~~~~~~~~~~
+- Update circuit validation for the current BBP sonata spec
+
 Version v0.12.1
 ---------------